Move ParentOperator to individual.py to resolve import conflict
