+++ fedot/core/optimisers/gp_comp/operators/crossover.py
@@ -5,9 +5,8 @@ from typing import TYPE_CHECKING, Any, Callable, List, Union
 from fedot.core.composer.constraint import constraint_function
 from fedot.core.log import Log
 from fedot.core.optimisers.gp_comp.gp_operators import equivalent_subtree, replace_subtrees
-from fedot.core.optimisers.gp_comp.individual import Individual
+from fedot.core.optimisers.gp_comp.individual import Individual, ParentOperator
 from fedot.core.optimisers.graph import OptGraph
-from fedot.core.optimisers.opt_history import ParentOperator
 from fedot.core.utilities.data_structures import ComparableEnum as Enum
 
 if TYPE_CHECKING:
